// src/entities/Country/model/types/country.ts
export enum Country {
  Russia = 'Russia',
  USA = 'USA',
  Germany = 'Germany',
  France = 'France',
  China = 'China',
  Japan = 'Japan',
  Italy = 'Italy',
  Spain = 'Spain',
  Canada = 'Canada',
  Brazil = 'Brazil',
  Australia = 'Australia',
  India = 'India',
  Mexico = 'Mexico',
  Argentina = 'Argentina',
  Turkey = 'Turkey',
  Switzerland = 'Switzerland',
}